arm: memoize NVIC pending scan; drop 128-bit divides from time conversion
Two hot-path costs measured by profile after the getenv latching:

1. arm_nvic_check_pending was ~14% of wall time.  arm_step re-checks
   pending interrupts every instruction while has_pending is set — and
   it stays set for long stretches when an exception is pending but
   MASKED (BASEPRI critical sections; a running handler that can't be
   preempted — TSCH executes whole slots inside the rtimer ISR).  Each
   re-check re-scanned all 8 ISPR words plus priorities.  Memoize the
   scan's winning (exception, priority) pair and invalidate only when
   pending/enable/priority state mutates (ISER/ICER/ISPR/ICPR/IPR/SHPR
   writes, ICSR pend set/clear, set/clear_pending, SysTick pend, and
   consuming the exception).  The per-instruction path becomes three
   integer compares.

2. cpu_ns_to_cycles / cpu_cycles_to_ns compiled to a __divti3 libcall
   (~3% of wall time; they run per scheduled event and per peripheral
   counter read).  Compute floor((a*b)/c) via the exact split identity
   (a/c)*b + ((a%c)*b)/c in pure 64-bit — bit-identical results, so
   event timing and timelines are unchanged (TSCH regression tests pass
   with identical simulated timestamps).  Same treatment for the two
   remaining __int128 divides in cc2538_sleeptimer.c, whose products
   provably fit int64.

Measured: nRF52840 TSCH 11.3x -> 14.2x realtime (94 -> 119 MIPS);
cc2538 RPL-UDP 265x -> 297x; nrf52840 RPL-UDP 342x -> 360x.

Regressions green: arm-correctness 153/153, msp430 correctness +
firmware, radio-medium 241/241, cc1200 73/73, both TSCH held-sync
tests (bit-identical timing), RPL-UDP dk/dongle/cc2538/sky, Zephyr
echo 120, nrf54l15 FLPR.

+++ src/arm/cc2538_sleeptimer.c
@@ -49,9 +49,10 @@ static void sleeptimer_schedule_compare(cc2538_sleeptimer_t *st) {
      * floor the event fired up to one tick early, so an ISR reading the
      * counter saw compare-1 (30.5 us early) — poison for TSCH slot
      * timing that immediately schedules the next deadline from "now". */
-    int64_t delta_cycles = (int64_t)(((__int128)delta * st->cpu->cpu_freq_hz
-                                      + SLEEPTIMER_FREQ_HZ - 1)
-                                     / SLEEPTIMER_FREQ_HZ);
+    /* delta <= 2^32, cpu_freq <= ~2^27 -> product < 2^59, fits int64. */
+    int64_t delta_cycles = ((int64_t)delta * st->cpu->cpu_freq_hz
+                            + SLEEPTIMER_FREQ_HZ - 1)
+                           / SLEEPTIMER_FREQ_HZ;
 
     arm_schedule_event(st->cpu, &st->compare_event,
                        st->cpu->cycles + delta_cycles);
@@ -84,8 +85,8 @@ static void sleeptimer_compare_fire(void *user_data, arm_event_t *ev) {
     if (elapsed > 0) {
         uint64_t ticks = (uint64_t)elapsed * SLEEPTIMER_FREQ_HZ / 1000000000ULL;
         st->base_count += (uint32_t)ticks;
-        st->base_ns += (int64_t)((__int128)ticks * 1000000000ULL
-                                 / SLEEPTIMER_FREQ_HZ);
+        /* ticks < 2^33 between fires -> ticks * 1e9 < 2^63, fits int64. */
+        st->base_ns += (int64_t)(ticks * 1000000000ULL / SLEEPTIMER_FREQ_HZ);
     }
 
     /* Set interrupt pending — wakes CPU from WFI */
